pyproject.toml
[tool.poetry]
name = "fastapi-oidc-backend"
version = "0.6.0"
description = "Build resource servers with FastAPI"
authors = ["Felix Gustavsson <felix@0b1.se>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/BlackVoid/fastapi-oidc-backend"
keywords = [ "fastapi", "authentication", "oidc", "openidconnect" ]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
]
packages = [
    { include = "*", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11.2"
fastapi = "^0"
httpx = "^0.27.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
black = "^24.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
